requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.3
